+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aave/react
 
+## 6.3.0
+
+### Minor Changes
+
+- 2dde326: **feat:** support multichain user claimable rewards queries
+
+### Patch Changes
+
+- Updated dependencies [2dde326]
+  - @aave/graphql@3.2.0
+  - @aave/client@6.3.0
+
 ## 6.2.1
 
 ### Patch Changes